Introduce STI for the various types of Account(s)
This introduces more flexibility for accounts.  Any transaction has a debit and a credit account.  However, these accounts can be different types:
* BankAccount
* CreditCardAccount
* TradingAccount - represents an organisation we interact with. Unlike the first two we don't keep track of a balance.

In order to support generic features associated with managing a balance, a Concern called Balances is mixed in for the first two.  We can check if an Account class has the Balance concern thru the class method #balances?

This commit includes suitable factories and specs.