Add support for both CSV with and without headers
Some financial institutions' CSV export includes a header file, others do not.  This change allows us to specify a column name or an index depending on which case is appropriate.

We have added a specific header attribute on ImportColumnsDefinition to flag which case we are dealing with.

To support this, we have changed the column attributes on ImportColumnsDefinition to the string type.  These are recast to Integer on reading, if the record shows that no header is given.

Various changes to specs and factories have been made to support this change.