Validates import column definitions
Adds validations to the import columns definition to ensure
required fields are present.

Specifically, it validates that either both credit and debit
columns are defined, or the amount column is defined, but not
a mix of the two. It also adds presence validations for date format,
date column and other party column.

Adds tests to confirm these validations.