Refactors file generation for transaction data
Creates a rake task to support generating test data for development.  It uses the same AccountTrxDataGenerator (previously LloydsImportFileGenerator) to generate the test data.

To support growing calls on AccountTrxDataGenerator, updated to allow transactions to
either be written to a CSV file, or to the database.

Updated tests to use the new generator functionality.