Adds transaction editing in place
Refactors transaction display and editing using a presenter.

This change allows users to edit transactions directly on the account show page, improving user experience.  Which fields are editable is a function of whether we are creating a new
transaction or editing an existing transaction.

It adds a TransactionPresenter to handle the display and form generation logic. It also modifies the routes and controller to support in-place editing and adds a new transaction button on the account page.